Added the possibility to customize testemonials
Besides customize testemonials, the solutions loading was
refactored so we could use a function in all code parts
that use this feature. (home and solution page)
