+++ index.php
@@ -1,6 +1,7 @@
 <?php 	
     $css_escolhido = 'index';
     require_once('header.php');
+    require_once('solucoes.php');
     $home = get_template_directory_uri(); 
 ?>
 <!-- Conteudo principal -->
@@ -46,61 +47,7 @@
         <div class="col-md-3"></div>
     </div>
     <!-- Carregamento das soluções cadastradas -->
-    <?php 
-    $args = array( 'post_type' => 'solucao', 'posts_per_page' => 3, 'order' => 'ASC' );            
-	$loop = new WP_Query( $args );
-	if( $loop->have_posts() ) { 
-        $counter = 0;
-        ?>
-	<?php while( $loop->have_posts()) {
-        // Flag para indicar se a solucao vai ser exibida na direita ou a esquerda
-        $alignRight = $counter%2 != 0;
-		$loop->the_post(); ?>
-        <div class="full-image-background" style="background-image: url(<?=the_post_thumbnail_url()?>)">
-            <div class="container vertical-center">
-                <div class="row mx-0">
-                    <?php if($alignRight) {
-                        $style_title = "right";
-                        ?>
-                        <div class="col-md-6 col-xs-4"></div>
-                    <?php
-                    } else {
-                        $style_title = "left";
-                    }
-                    ?>
-                    <div class="col-md-6 col-xs-8" align="<?=$style_title?>" >
-                        <h2 class="titulo-solucao mb-4"><?php the_title();?></h2>
-                    </div>
-                </div>
-                <div class="row mx-0">
-                    <?php if($alignRight) { ?>
-                        <div class="col-md-8 col-xs-8"></div>
-                    <?php
-                    }
-                    ?>
-                    <div class="col-md-4 col-xs-8 texto-solucao block-with-text mb-4">
-                        <?php the_content();?>
-                    </div>
-                </div>
-                <div class="row mx-0">
-                    <?php if($alignRight) { ?>
-                        <div class="col-md-10 col-xs-8"></div>
-                    <?php
-                    }
-                    ?>
-                    <div class="col-md-2 col-xs-4" align="center">
-                        <a href="<?= the_permalink();?>" style="text-decoration: none">
-                            <button type="button" class="btn btn-outline-secondary btn-block">Saiba mais</button>
-                        </a>
-                    </div>
-                </div>
-            </div>
-        </div>
-    <?php
-        $counter = $counter + 1;
-        } 
-    } 
-    ?>
+    <?php get_solucoes(3);?>
 </div>
 <div class="home-blog container">
 </div>
